\name{BSVGP-package}
\alias{BSVGP-package}
\alias{BSVGP}
\docType{package}
\title{
\packageTitle{BSVGP}
}
\description{
\packageDescription{BSVGP}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BSVGP}
\packageIndices{BSVGP}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{BSVGP}

Maintainer: \packageMaintainer{BSVGP}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
