\name{TCGP-package}
\alias{TCGP-package}
\alias{TCGP}
\docType{package}
\title{
\packageTitle{TCGP}
}
\description{
\packageDescription{TCGP}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{TCGP}
\packageIndices{TCGP}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{TCGP}

Maintainer: \packageMaintainer{TCGP}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
